+++ ios/Classes/UnityAdsConstants.swift
@@ -4,25 +4,34 @@ struct UnityAdsConstants {
     static let BANNER_AD_CHANNEL = MAIN_CHANNEL + "/bannerAd";
     static let VIDEO_AD_CHANNEL = MAIN_CHANNEL + "/videoAd";
     
+    static let PLACEMENT_ID_PARAMETER = "placementId";
+    static let ERROR_CODE_PARAMETER = "errorCode";
+    static let ERROR_MESSAGE_PARAMETER = "errorMessage";
+    
+    //initialize
+    static let INIT_METHOD = "init";
     static let GAME_ID_PARAMETER = "gameId";
     static let TEST_MODE_PARAMETER = "testMode";
-    static let FIREBASE_TEST_LAB_MODE_PARAMETER = "firebaseTestLabMode";
+    static let INIT_COMPLETE_METHOD = "initComplete";
+    static let INIT_FAILED_METHOD = "initFailed";
     
-    static let PLACEMENT_ID_PARAMETER = "placementId";
+    //load
+    static let LOAD_METHOD = "load";
+    static let LOAD_COMPLETE_METHOD = "loadComplete";
+    static let LOAD_FAILED_METHOD = "loadFailed";
+    
+    //show
+    static let SHOW_VIDEO_METHOD = "showVideo";
     static let SERVER_ID_PARAMETER = "serverId";
+    static let SHOW_COMPLETE_METHOD = "showComplete";
+    static let SHOW_FAILED_METHOD = "showFailed";
+    static let SHOW_START_METHOD = "showStart";
+    static let SHOW_SKIPPED_METHOD = "showSkipped";
+    static let SHOW_CLICK_METHOD = "showClick";
+    
+    //banner
     static let HEIGHT_PARAMETER = "height";
     static let WIDTH_PARAMETER = "width";
-    
-    static let INIT_METHOD = "init";
-    static let IS_READY_METHOD = "isReady";
-    static let SHOW_VIDEO_METHOD = "showVideo";
-    
-    static let READY_METHOD = "ready";
-    static let START_METHOD = "start";
-    static let COMPLETE_METHOD = "complete";
-    static let SKIPPED_METHOD = "skipped";
-    static let ERROR_METHOD = "error";
-    
     static let BANNER_ERROR_METHOD = "banner_error";
     static let BANNER_LOADED_METHOD = "banner_loaded";
     static let BANNER_CLICKED_METHOD = "banner_clicked";